requirements.txt
flask
transformers
scipy
requests
python-dotenv
torch
